requirements.txt
amqp==2.6.1
ansible-vault==2.1.0
appdirs==1.4.4
asgiref==3.3.4
bcrypt==3.1.7
billiard==3.6.3.0
boto3==1.17.68
botocore==1.20.68
celery==4.4.7
certifi==2020.6.20
cffi==1.14.3
channels==3.0.4
channels-redis==3.4.0
chardet==3.0.4
croniter==1.3.4
cryptography==36.0.1
distlib==0.3.1
Django==3.2.13
django-celery-beat==2.0.0
django-storages==1.11.1
django-timezone-field==4.0
django-unixdatetimefield==1.2.0
django-utils-six==2.0 
djangorestframework==3.11.2
filelock==3.0.12
fsspec==2021.4.0
geoip2==2.9.0
hvac==0.10.14
idna==2.8
importlib-metadata==2.0.0
jmespath==0.10.0
kombu==4.6.11
loguru==0.5.3
maxminddb==2.0.3
numpy==1.21.0
openpyxl==3.0.7
pandas==1.2.4
pika==1.2.0
psycopg2-binary==2.8.6
pycparser==2.20
pycryptodome==3.10.1
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
pymemcache==3.5.0
pytz==2021.1
requests==2.25.1
s3fs==0.4.2
s3transfer==0.4.2
six==1.16.0
sqlparse==0.3.1
urllib3==1.26.5
uWSGI==2.0.19.1
validate-docbr==1.10.0 
vine==1.3.0
virtualenv==20.0.33
XlsxWriter==1.4.3
zipp==3.3.0
xlrd >= 1.0.0
